+++ src/views/Badge.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
 import { ref } from 'vue'
+import { ClockCircleOutlined, MinusOutlined, PlusOutlined } from '@ant-design/icons-vue'
 const value = ref(5)
 const dot = ref(true)
 const colors = [
@@ -39,23 +40,7 @@ function increase() {
       </Badge>
       <Badge>
         <template #value>
-          <svg
-            focusable="false"
-            class="u-svg"
-            data-icon="clock-circle"
-            width="1em"
-            height="1em"
-            fill="currentColor"
-            aria-hidden="true"
-            viewBox="64 64 896 896"
-          >
-            <path
-              d="M512 64C264.6 64 64 264.6 64 512s200.6 448 448 448 448-200.6 448-448S759.4 64 512 64zm0 820c-205.4 0-372-166.6-372-372s166.6-372 372-372 372 166.6 372 372-166.6 372-372 372z"
-            ></path>
-            <path
-              d="M686.7 638.6L544.1 535.5V288c0-4.4-3.6-8-8-8H488c-4.4 0-8 3.6-8 8v275.4c0 2.6 1.2 5 3.3 6.5l165.4 120.6c3.6 2.6 8.6 1.8 11.2-1.7l28.6-39c2.6-3.7 1.8-8.7-1.8-11.2z"
-            ></path>
-          </svg>
+          <ClockCircleOutlined style="color: #f5222d" />
         </template>
         <Avatar shape="square" size="large" />
       </Badge>
@@ -151,33 +136,10 @@ function increase() {
           <Avatar shape="square" size="large" />
         </Badge>
         <Button @click="decline">
-          <svg
-            class="u-icon"
-            focusable="false"
-            data-icon="minus"
-            width="1em"
-            height="1em"
-            fill="currentColor"
-            aria-hidden="true"
-            viewBox="64 64 896 896"
-          >
-            <path d="M872 474H152c-4.4 0-8 3.6-8 8v60c0 4.4 3.6 8 8 8h720c4.4 0 8-3.6 8-8v-60c0-4.4-3.6-8-8-8z"></path>
-          </svg>
+          <MinusOutlined />
         </Button>
         <Button @click="increase">
-          <svg
-            class="u-icon"
-            focusable="false"
-            data-icon="plus"
-            width="1em"
-            height="1em"
-            fill="currentColor"
-            aria-hidden="true"
-            viewBox="64 64 896 896"
-          >
-            <path d="M482 152h60q8 0 8 8v704q0 8-8 8h-60q-8 0-8-8V160q0-8 8-8z"></path>
-            <path d="M192 474h672q8 0 8 8v60q0 8-8 8H160q-8 0-8-8v-60q0-8 8-8z"></path>
-          </svg>
+          <PlusOutlined />
         </Button>
       </Space>
       <Space gap="large" align="center">
@@ -214,11 +176,5 @@ function increase() {
   color: #fff;
   border-radius: 10px;
   box-shadow: 0 0 0 1px #ffffff;
-}
-.u-svg {
-  fill: #f5222d;
-}
-.u-icon {
-  fill: rgba(0, 0, 0, 0.88);
 }
 </style>